raco pkg {install,update}: add `--clone <dir>` mode
Using `--clone <dir>` with a Git-based package source causes the
package installation to be linked to a clone of the repository
as a subdirectory of <dir>. The package can be developed locally
in the usual way with Git tools, but `raco pkg update` can itself
pull updates to the package/repository.

See the new chapter 6 in "Package Management in Racket" for
more information.

original commit: 6379aaddef7fa26d14ab2523efa8aa5492f6ac2b